requirements.txt
faiss-cpu>=1.8.0
numpy>=1.24.0
tqdm>=4.65.0
pydantic>=2.0.0
docx-parser>=0.1.0
langchain>=0.3.0
langchain-community>=0.3.0
pytest>=7.0.0
